pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "myelin8"
version = "2.0.0a1"
description = "Brain-inspired AI memory compression with optional post-quantum encryption. Extends context windows 3-50x through tiered storage and semantic indexing."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Kevin Qi", email = "kevin@qinnovate.com"},
]
keywords = [
    "ai", "memory", "compression", "encryption", "post-quantum",
    "context-window", "tiered-storage", "zstd", "parquet", "ml-kem-768",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: System :: Archiving :: Compression",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "zstandard>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
frozen = [
    "pyarrow>=14.0.0",
]
embeddings = [
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "pyarrow>=14.0.0",
]
governance = [
    "pyyaml>=6.0",
    "jsonschema>=4.17",
]

[project.urls]
Homepage = "https://github.com/qinnovates/myelin8"
Repository = "https://github.com/qinnovates/myelin8"
Issues = "https://github.com/qinnovates/myelin8/issues"
Documentation = "https://github.com/qinnovates/myelin8#readme"

[project.scripts]
myelin8 = "src.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
where = ["."]
